Shared settings moved to base.py
