+++ src/learning/admin/__init__.py
@@ -1,8 +1,4 @@
 from django.conf import settings
-from .actions import (
-    confirm_action,
-    disprove_action
-)
 
 if settings.DEBUG:
     from .development import *
